separated components from App.js
