+++ web/film-club-client/src/components/FilmDetailPage.js
@@ -1,0 +1,16 @@
+
+
+//////  DETAIL PAGE (WIP) //////
+// const detailPage = (props) => {
+//   return (
+//     <Box >
+//       <Image src={`http://localhost:8080/v1/films/images${props.movie.posterPath}`}></Image>
+//       <Box>
+//         <Heading level="2">{props.movie.title}</Heading>
+//         <br/>
+//         <Text>{props.movie.overview}</Text>
+//       </Box>
+//     </Box>
+//   )
+// }
+////// ////// //////
