+++ web/film-club-client/src/components/Sidebar.js
@@ -1,0 +1,43 @@
+    
+    
+    // <Box direction="row" flex overflow={{ horizontal: 'hidden' }} pad={{top:"20%", bottom: "5%"}}>
+    // <Box flex align="center" justify="center">
+    //     {movies.map((movie, id) => (
+    //     <PosterCard key={id} movie={movie} />
+    //     ))}
+    // </Box>
+    // {(!showSidebar || size !== 'small') ? (
+    //     <Collapsible direction="horizontal" open={showSidebar}>
+    //     <Box
+    //         flex
+    //         width="medium"
+    //         height="small"
+    //         background="light-2"
+    //         elevation="small"
+    //         align="center"
+    //         justify="center"
+    //     >
+    //         menu items here
+    //     </Box>
+    //     </Collapsible>
+    // ) : (
+    //     <Sidebar width="medium" collapsible="true">
+    //         <Box
+    //         background="light"
+    //         tag="header"
+    //         justify="end"
+    //         align="center"
+    //         direction="row"
+    //         >
+    //         <Button icon={<FormClose />} onClick={() => setShowSidebar(false)} />
+    //         </Box>
+    //         <Box
+    //         background='light-2'
+    //         align="center"
+    //         justify="center"
+    //         >
+    //             menu items here
+    //         </Box>  
+    //     </Sidebar>
+    // )}
+    // </Box>
